requirements.txt
attrs==19.3.0
# llvmlite==0.31.0
multidict==4.7.4
multimethod==1.2
pytest==5.3.5
mypy-extensions==0.4.3
typing-extensions==3.7.4.1
typing-inspect==0.5.0
